[monodroid] Use managed API to load assemblies in-memory
We ideally want to have the same loading semantic than `Assembly.Load(byte[])` which ensures that a new MonoAssembly* instance is always created per-appdomain (MONO_ASMCTX_INDIVIDUAL context kind in Mono).

Unfortunately there is no public unmanaged API to achieve this, the only way to pass MONO_ASMCTX_INDIVIDUAL is via the MonoAssemblyLoadRequest functions which are not accessible.

Thus this commit switches to using the managed API equivalent (a technique we already use to create appdomains) which does incur a performance penalty (two extra array copying operations and an extra native->managed->native roundtrip) but is more accurate.