[monodroid] Fix issue with register and in-memory loaded assemblies
Previously when we were loading Mono.Android from disk, there would usually only be one version of it used and shared across assemblies.

When switching over to supply assemblies directly from in-memory content, the following error started popping up:

```
Unhandled Exception:
 System.ArgumentException: Handle value is not valid.
 Parameter name: method
  at Java.Interop.JniEnvironment+InstanceMethods.CallObjectMethod (Java.Interop.JniObjectReference instance, Java.Interop.JniMethodInfo method) [0x0002f] in <2763e1c7356b4103af4f02faed41bb5c>:0
  at Android.Runtime.JNIEnv.CallObjectMethod (System.IntPtr jobject, System.IntPtr jmethod) [0x0000e] in <70c334f9ebd34e718247cb9f549396f0>:0
  at Java.Interop.TypeManager.GetClassName (System.IntPtr class_ptr) [0x00000] in <70c334f9ebd34e718247cb9f549396f0>:0
  at Android.Runtime.JNIEnv.RegisterJniNatives (System.IntPtr typeName_ptr, System.Int32 typeName_len, System.IntPtr jniClass, System.IntPtr methods_ptr, System.Int32 methods_len) [0x000d3] in <70c334f9ebd34e718247cb9f549396f0>:0
```

This strange error actually stems from the fact that it appears to be running in an appdomain context where the static JNIEnv has not been initialized (thus the static field `mid_Class_getName` is IntPtr.Zero).

This was puzzling because we do initialize the runtime correctly when creating a new appdomain, however after investigating further the reason turned out to be that `RegisterJniNatives` was indeed being called via a method handle that came from a JNIEnv's MonoClass that was not in the proper domain thus causing the mismatch.

The previous code would rely on the fact that on desktop the `registerType` static global variable would be overwritten everytime the android runtime was initialized. This is wrong because, for the same reason than we reset the TypeManager every time we change active context, the MonoMethod representing RegisterJniNatives also needs to be changed to point at the version that's loaded in the current domain.

The reason that it broke now is because loading the assembly from disk was actually working around the problem previously since the assembly would end up being shared across domains thus the registerType method handle would work fine no matter which domain it was being executed in. Switching to in-memory loading and thus individual assemblies revealed the problem.